refactored

+++ s3_length_attribute.py
@@ -2,7 +2,7 @@
 Tasks
 1. Calculate length from shape nodes file
 """
-
+import os
 import pandas as pd
 import numpy as np
 import geopandas as gpd
@@ -34,17 +34,22 @@ def stringtolist(string):
         d+=d_
     return (d*1000)
 
+if not os.path.isfile('../midstages/shapenodes_unidirec.csv'):
+    cols_interest = ['LINK_ID', 'ST_NAME','REF_IN_ID', 'NREF_IN_ID', 'N_SHAPEPNT', 'FUNC_CLASS', 'SPEED_CAT','LANE_CAT','DIR_TRAVEL','PHYS_LANES','FROM_LANES','TO_LANES','geometry']
+    all_links_ca = pd.read_csv('../midstages/all_links.csv', usecols = cols_interest)
+    b1 = all_links_ca[all_links_ca['DIR_TRAVEL'] =='B']
 
-# Preprocessing shape nodes file to make it unidirectional links. Length is calculated fater that
-sn = pd.read_csv('nav_street_for_mobiliti_links_shapenodes.csv', names = ['global_id','LINK_ID','nodes']) #shape nodes from Jane
-#create additional links in shape nodes column
-sn_bi = sn[sn['LINK_ID'].isin(b1['LINK_ID'].values)] #bidirec
-s2 = sn_bi.copy()
-s2['LINK_ID'] = np.arange(7000000000,7000000000+len(sn_bi))
-sn_FT = sn[~sn['LINK_ID'].isin(b1['LINK_ID'].values)]  #unidirec FT links
-shape_nodes_unidirec = sn_FT.append([s2,sn_bi])
-shape_nodes_unidirec = shape_nodes_unidirec.drop('global_id', axis = 1)
-shape_nodes_unidirec.to_csv('../midstages/shapenodes_unidirec.csv', index = False)
+    # Preprocessing shape nodes file to make it unidirectional links. Length is calculated fater that
+    sn = pd.read_csv('../midstages/nav_street_for_mobiliti_links_shapenodes.csv', names = ['global_id','LINK_ID','nodes']) #shape nodes from Jane
+    #create additional links in shape nodes column
+    sn_bi = sn[sn['LINK_ID'].isin(b1['LINK_ID'].values)] #bidirec
+    s2 = sn_bi.copy()
+    s2['LINK_ID'] = np.arange(7000000000,7000000000+len(sn_bi))
+    sn_FT = sn[~sn['LINK_ID'].isin(b1['LINK_ID'].values)]  #unidirec FT links
+    shape_nodes_unidirec = sn_FT.append([s2,sn_bi])
+    shape_nodes_unidirec = shape_nodes_unidirec.drop('global_id', axis = 1)
+    shape_nodes_unidirec.to_csv('../midstages/shapenodes_unidirec.csv', index = False)
+
 
 # Calculating length from the unidirectional shapenodes file
 shape_nodes = pd.read_csv('../midstages/shapenodes_unidirec.csv')
